--- core/src/main/java/dev/kscott/interfaces/core/element/Element.java
@@ -1,8 +1,0 @@
-package dev.kscott.interfaces.core.element;
-
-/**
- * Represents an element that can be rendered to a pane.
- */
-public interface Element {
-
-}
