--- core/src/main/java/dev/kscott/interfaces/core/view/InterfaceViewer.java
@@ -1,17 +1,0 @@
-package dev.kscott.interfaces.core.view;
-
-import org.checkerframework.checker.nullness.qual.NonNull;
-
-/**
- * Represents a target that can view an interface.
- */
-public interface InterfaceViewer {
-
-    /**
-     * Displays a view to the viewer.
-     *
-     * @param pane the pane
-     */
-    void open(@NonNull InterfaceView<?, ?, ?> pane);
-
-}
