+++ core/src/main/java/org/incendo/interfaces/core/element/Element.java
@@ -1,0 +1,8 @@
+package org.incendo.interfaces.core.element;
+
+/**
+ * Represents an element that can be rendered to a pane.
+ */
+public interface Element {
+
+}
